Move `TxCBOR` to `Cardano.Wallet.Read.Tx.CBOR`
