+++ lib/read/lib/Cardano/Read/Ledger/Tx/CBOR.hs
@@ -1,27 +1,17 @@
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
-{-# LANGUAGE TypeOperators #-}
 
-{-# OPTIONS_GHC -Wno-orphans #-}
-{-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
+{- |
+Copyright: © 2020-2022 IOHK, 2024 Cardano Foundation
+License: Apache-2.0
 
--- |
--- Copyright: © 2020-2022 IOHK
--- License: Apache-2.0
---
--- CBOR operations for era dependent transactions.
---
-
+Binary serialization of transactions.
+-}
 module Cardano.Read.Ledger.Tx.CBOR
-    ( TxCBOR
-    , renderTxToCBOR
-    , parseTxFromCBOR
-    , serializeTx
+    ( serializeTx
     , deserializeTx
-    , roundTripTxCBor
     )
     where
 
@@ -46,33 +36,13 @@ import Cardano.Read.Ledger.Eras
     , IsEra (..)
     )
 import Cardano.Wallet.Read.Eras
-    ( EraValue
-    , K (..)
-    , applyEraFunValue
-    , extractEraValue
-    , sequenceEraValue
+    ( K (..)
     , unK
     , (:.:) (..)
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
     , TxT
-    )
-import Data.ByteArray.Encoding
-    ( Base (Base16)
-    , convertToBase
-    )
-import Data.ByteString.Lazy
-    ( toStrict
-    )
-import Data.Text.Class
-    ( ToText
-    )
-import Data.Text.Encoding
-    ( decodeUtf8
-    )
-import Fmt
-    ( Buildable (..)
     )
 import Ouroboros.Consensus.Shelley.Eras
     ( StandardAllegra
@@ -86,20 +56,6 @@ import Ouroboros.Consensus.Shelley.Eras
 import qualified Cardano.Ledger.Binary.Encoding as Ledger
 import qualified Data.ByteString.Lazy as BL
 
--- | Serialized version of a transaction. Deserializing should at least expose
--- enough information to compute the `TxId`.
-type TxCBOR = EraValue (K BL.ByteString)
-
-instance Buildable TxCBOR where
-    build
-        = build . decodeUtf8 . convertToBase Base16 . toStrict . extractEraValue
-
-instance ToText TxCBOR
-
--- | Render a tx into its cbor, it just applies 'serializeTx'.
-renderTxToCBOR :: EraValue Tx -> EraValue (K BL.ByteString)
-renderTxToCBOR = applyEraFunValue serializeTx
-
 {-# INLINABLE serializeTx #-}
 -- | CBOR serialization of a tx in any era.
 serializeTx :: forall era . IsEra era => Tx era -> K BL.ByteString era
@@ -111,20 +67,15 @@ serializeTx = case theEra @era of
     Alonzo -> f (eraProtVerLow @StandardAlonzo)
     Babbage -> f (eraProtVerLow @StandardBabbage)
     Conway -> f (eraProtVerLow @StandardConway)
-
   where
     f :: EncCBOR (TxT era) => Ledger.Version -> Tx era -> K BL.ByteString era
     f protVer = K . Ledger.serialize protVer . unTx
 
--- | Parse CBOR into a transaction in any eras
--- , smart application  of `deserializeTx`.
-parseTxFromCBOR :: TxCBOR -> Either DecoderError (EraValue Tx)
-parseTxFromCBOR = sequenceEraValue
-    . applyEraFunValue deserializeTx
-
 {-# INLINABLE deserializeTx #-}
 -- | CBOR deserialization of a tx in any era.
-deserializeTx :: forall era . IsEra era => K BL.ByteString era -> (Either DecoderError :.: Tx) era
+deserializeTx
+    :: forall era . IsEra era
+    => K BL.ByteString era -> (Either DecoderError :.: Tx) era
 deserializeTx = case theEra @era of
     Byron -> \txCBOR ->
         Comp $ Tx <$> decodeFull byronProtVer (unK txCBOR)
@@ -134,9 +85,6 @@ deserializeTx = case theEra @era of
     Alonzo -> decodeTx (eraProtVerLow @StandardAlonzo) "AlonzoTx"
     Babbage -> decodeTx (eraProtVerLow @StandardBabbage) "BabbageTx"
     Conway -> decodeTx (eraProtVerLow @StandardConway) "ConwayTx"
-    where
+  where
     decodeTx protVer label (K txCBOR) =
         Comp $ Tx <$> decodeFullAnnotator protVer label decCBOR txCBOR
-
-roundTripTxCBor :: TxCBOR -> Either DecoderError TxCBOR
-roundTripTxCBor = fmap renderTxToCBOR . parseTxFromCBOR
