+++ lib/unit/test/unit/Cardano/Wallet/DB/Arbitrary.hs
@@ -49,9 +49,6 @@ import Cardano.Mnemonic
 import Cardano.Pool.Types
     ( PoolId (..)
     )
-import Cardano.Read.Ledger.Tx.CBOR
-    ( TxCBOR
-    )
 import Cardano.Wallet.Address.Book
     ( AddressBookIso (..)
     )
@@ -220,6 +217,9 @@ import Cardano.Wallet.Primitive.Types.UTxO
 import Cardano.Wallet.Read.Eras
     ( eraValueSerialize
     , knownEraIndices
+    )
+import Cardano.Wallet.Read.Tx.CBOR
+    ( TxCBOR
     )
 import Cardano.Wallet.Unsafe
     ( someDummyMnemonic
