+++ lib/unit/test/unit/Cardano/Wallet/DB/Store/Transactions/StoreSpec.hs
@@ -21,9 +21,6 @@ import Cardano.DB.Sqlite
     ( ForeignKeysSetting (..)
     , runQuery
     )
-import Cardano.Read.Ledger.Tx.CBOR
-    ( roundTripTxCBor
-    )
 import Cardano.Wallet.DB.Arbitrary
     ()
 import Cardano.Wallet.DB.Fixtures
@@ -62,6 +59,9 @@ import Cardano.Wallet.DB.Store.Transactions.TransactionInfo
     )
 import Cardano.Wallet.Primitive.Types.Tx
     ( Tx (..)
+    )
+import Cardano.Wallet.Read.Tx.CBOR
+    ( roundTripTxCBor
     )
 import Control.Monad
     ( forM_
