+++ lib/wallet/src/Cardano/Wallet/DB/Store/Transactions/TransactionInfo.hs
@@ -12,10 +12,6 @@ module Cardano.Wallet.DB.Store.Transactions.TransactionInfo
 
 import Prelude
 
-import Cardano.Read.Ledger.Tx.CBOR
-    ( TxCBOR
-    , renderTxToCBOR
-    )
 import Cardano.Read.Ledger.Tx.CollateralInputs
     ( getEraCollateralInputs
     )
@@ -116,6 +112,10 @@ import Cardano.Wallet.Read.Eras
     ( EraValue
     , IsEra
     , applyEraFun
+    )
+import Cardano.Wallet.Read.Tx.CBOR
+    ( TxCBOR
+    , renderTxToCBOR
     )
 import Cardano.Wallet.Transaction
     ( ValidityIntervalExplicit (invalidHereafter)
