+++ lib/wallet/src/Cardano/Wallet/Primitive/Types/Tx/TransactionInfo.hs
@@ -24,9 +24,6 @@ module Cardano.Wallet.Primitive.Types.Tx.TransactionInfo
 
 import Prelude
 
-import Cardano.Read.Ledger.Tx.CBOR
-    ( TxCBOR
-    )
 import Cardano.Wallet.Primitive.Types.Coin
     ( Coin (..)
     )
@@ -51,6 +48,9 @@ import Cardano.Wallet.Primitive.Types.Tx.TxMeta
     )
 import Cardano.Wallet.Primitive.Types.Tx.TxOut
     ( TxOut
+    )
+import Cardano.Wallet.Read.Tx.CBOR
+    ( TxCBOR
     )
 import Control.DeepSeq
     ( NFData (..)
